MultiRoleFilterProxzModel class in a different file
